// src/app/interfaces/inmobiliario.ts
//Se crea una interface para los datos traidos de la base de datos
export interface Inmobiliario {
    id_propiedad?: string;
    imagen: string;
    areatotal: string;
    ncuartos: string;
    nbanos: string;
    nparqueaderos: string;
    valor_propiedad: string;
}

/* Created By: Juan Felipe Rodríguez Rendón */